requirements.txt
bchlib==0.14.0
torch==1.6.0
torchvision==0.7.0
tensorflow-gpu==1.15